Docker ML/ML models with Dockercompose & unit_testing/app/requirements.txt
Flask==2.3.3
scikit-learn==1.3.0
joblib==1.3.2
numpy==1.24.3
pandas==2.0.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0